restructured hierarchy
