+++ frontend/src/modules/customers/components/auth/CustomerLogIn.tsx
@@ -1,14 +1,26 @@
 import { Box, Button, Link, Stack, Typography } from "@mui/material";
 import { useFormik } from "formik";
-import TextInput from "@/shared/components/inputs/TextInput";
-import { useAuthService } from "../hooks/useAuthService";
+import TextInput from "@/shared/inputs/TextInput";
+import { useAuthService } from "../../../../shared/hooks/useAuthService";
 import { useRouter } from "next/router";
-import {
-  LogInFormType,
-  logInInitialValues,
-  logInValidationSchema,
-} from "./config";
 import toast from "react-hot-toast";
+import * as yup from "yup";
+import { Schema } from "yup";
+
+export interface LogInFormType {
+  email: string;
+  password: string;
+}
+
+export const logInValidationSchema: Schema<LogInFormType> = yup.object().shape({
+  email: yup.string().required("Email is required").email("Invalid email"),
+  password: yup.string().required("Password is required"),
+});
+
+export const logInInitialValues: LogInFormType = {
+  email: "",
+  password: "",
+};
 
 const CustomerLogIn = () => {
   const authService = useAuthService();
