+++ frontend/src/modules/customers/components/menu/CustomerMenu.tsx
@@ -1,9 +1,9 @@
 import { useMenuService } from "@/shared/hooks/useMenuService";
-import { Box, Divider, Stack, Typography } from "@mui/material";
+import { Box, Stack, Typography } from "@mui/material";
 import { useEffect, useState } from "react";
-import ComboCard from "../order/components/ComboCard";
-import { ActiveMenuResponse } from "../../../../../backend/src/menu/menu.dto";
-import { ItemCard } from "../order/components/ItemCard";
+import ComboCard from "./ComboCard";
+import { ActiveMenuResponse } from "../../../../../../backend/src/menu/menu.dto";
+import { ItemCard } from "./ItemCard";
 
 const CustomerMenu = ({ isOrderPage }: { isOrderPage: boolean }) => {
   const { getActiveMenu } = useMenuService();
