+++ frontend/src/pages/customers/profile/index.tsx
@@ -23,10 +23,10 @@ import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 import WarningIcon from "@mui/icons-material/Warning";
 import { isFlaggedAccount } from "../../../shared/components/navbar";
 import { useRouter } from "next/router";
-import { useUserService } from "@/modules/customers/user/hooks/useUserService";
+import { useUserService } from "@/shared/hooks/useUserService";
 import { useEffect, useState } from "react";
 import { UserProfile } from "../../../../../backend/src/user/user.dto";
-import { useAuthService } from "@/modules/customers/auth/hooks/useAuthService";
+import { useAuthService } from "@/shared/hooks/useAuthService";
 import toast from "react-hot-toast";
 import BaseDialog from "@/shared/components/baseDialog";
 import DropdownMenu from "@/shared/components/DropdownMenu";
