+++ frontend/src/shared/hooks/useAuthService.ts
@@ -1,6 +1,6 @@
 import { useBaseAPI } from "@/shared/hooks/useBaseAPI";
-import { SignUpDto } from "../../../../../../backend/src/auth/auth.dto";
-import { LogInDto } from "../../../../../../backend/src/auth/auth.dto";
+import { SignUpDto } from "../../../../backend/src/auth/auth.dto";
+import { LogInDto } from "../../../../backend/src/auth/auth.dto";
 
 export const useAuthService = () => {
   const authControllerBase = "auth";
